requirements.txt
python-telegram-bot[webhooks]
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
flask
flask-cors
pytz
requests
python-dateutil
APScheduler
